Refactor link API requests
