+++ client/src/API/API.ts
@@ -1,75 +1,13 @@
-import { Document, LinkType } from "../utils/interfaces";
 import { documentAPI } from "./documentAPI";
 import { authAPI } from "./authAPI";
+import { linkAPI } from "./linkAPI";
 
 export const baseURL = "http://localhost:3000";
-
-/*************************   DOCUMENTS   *****************************/
-
-/****************************   LINKS   **********************************/
-
-async function getLinks(id: number) {
-  try {
-    const response = await fetch(baseURL + `/${id}/links`);
-    if (response.ok) {
-      const links = await response.json();
-      return links;
-    }
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-async function putLink(
-  targetDocumentId: number,
-  linkType: LinkType[],
-  id: number
-) {
-  try {
-    const response = await fetch(baseURL + `/${id}/links`, {
-      method: "PUT",
-      credentials: "include",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ targetDocumentId, linkType }),
-    });
-    if (response.ok) {
-      return true;
-    }
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-async function deleteLink(
-  sourceNodeId: number,
-  targetNodeId: number,
-  linkType: string
-) {
-  try {
-    const response = await fetch(baseURL + `/links`, {
-      method: "DELETE",
-      credentials: "include",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ sourceNodeId, targetNodeId, linkType }),
-    });
-    if (response.ok) {
-      return true;
-    }
-  } catch (err) {
-    console.error(err);
-  }
-}
 
 const API = {
   ...authAPI,
   ...documentAPI,
-  getLinks,
-  putLink,
-  deleteLink,
+  ...linkAPI,
 };
 
 export default API;
